moved temp checker to analysis
