+++ ipsuite/analysis/__init__.py
@@ -13,6 +13,7 @@ from ipsuite.analysis.model import (
     ForceDecomposition,
     MDStability,
     NaNCheck,
+    TemperatureCheck,
     Prediction,
     PredictionMetrics,
     RattleAtoms,
@@ -35,6 +36,7 @@ __all__ = [
     "BoxScale",
     "BoxHeatUp",
     "NaNCheck",
+    "TemperatureCheck",
     "ConnectivityCheck",
     "EnergySpikeCheck",
     "MDStability",
